Update cloud-sdk to 276
app engine SDK is not being updated regularly but we need to keep up
with gcloud sdk versions somehow. Bumping the version of cloud-sdk to
force a new build.

According to [this doc][old] the standalone app engine SDK will not be
available to download as of 2020 July 30. It has been deprecated as of
2019 July 30.

[old]: https://cloud.google.com/appengine/docs/standard/python/sdk-gcloud-migration